packet system refactor

+++ SR2MP/Components/Time/NetworkTime.cs
@@ -28,7 +28,7 @@ public sealed class NetworkTime : MonoBehaviour
 
         var packet = new WorldTimePacket
         {
-            Type = (byte)PacketType.WorldTime,
+            Type = PacketType.WorldTime,
             Time = timeDirector._worldModel.worldTime
         };
